packages for real-time collaboration in JupyterLab
JupyterLab was on major version 2 before, apparently relating to
how dependencies were resolved (it looks like that permitted later
versions of one or more other packages to be installed that
conda/mamba prefers to upgrade, perhaps because they are more
likely to appear as dependencies of other packages). This takes it
to version 3. I've specified 3.1 as the minimum since that's when
real-time collaboration was first introduced:

    https://jupyterlab.readthedocs.io/en/stable/user/rtc.html

But conda/mamba will install the latest version.

See that linked page for further information on RTC in JupyterLab.

I have not observed changes since version 2, other than RTC, that
seem likely to affect our immediate use. In the longer term, using
the newest (stable) version is probably best.

I've kept IPython back at major version 7, just for now, to avoid
taking black (https://github.com/psf/black) as a dependency, so we
don't have to worry about editors/IDEs automatically detecting
black and using it for autoformatting.