take IPython 8.0.1 to fix CVE-2022-21699
Detailed info:

https://ipython.readthedocs.io/en/stable/whatsnew/version8.html#ipython-8-0-1-cve-2022-21699

I could've kept at major version 7 but I figured this is easier.
Since the last commit, I've checked if VS Code automatically uses
or suggests using black to autoformat because it is installed in
the repository as a dependency of IPython >=8.0.0, and it does not
(at least in the configuration we're using), so I'm not too worried
about that.

Autoformatting in an IPython shell run from a terminal with the
ipython command automatically (and destructively) rewrites input,
which is undesirable for our purposes, but we've never been using
IPython that way, so I'm not worried about that. We can reconfigure
it if necessary.