removed accidentally added import statement that broke decorators.py
Our decorators module had an accidentally added import statement (a
pair of spurious keystrokes will do this in the editor) in commit
b54f3f7 that caused a ModuleNotFoundError to be raised, breaking
all uses of our decorators module including searching for and
running doctests. The error messages looked like this:

decorators.py:4: in <module>
    from curses import wrapper
C:\Users\ek\miniforge3\envs\algoviz-basics\lib\curses\__init__.py:13: in <module>
    from _curses import *
E   ModuleNotFoundError: No module named '_curses'

The editor apparently offered to import wrapper from the curses
module, presumably when opening the "lightbulb" menu with
Ctrl+comma while writing "wrapper" as a new identifier. The curses
module is part of Python's standard library, but it is only
actually usable on Unix-like systems, since those are the systems
for which libraries like curses and ncurses exist (and tend to be
present).

Thus, on Windows, the C module standing behind the curses module,
which is called _curses, could not be found, because distributions
of Python for non-Unix operating systems don't have it (because
it wouldn't link on those systems, there being no curses library
for it to link to, and because it wouldn't likely compile on them
either, there being no curses library headers for its #include
directives to resolve).

Attempting to import wrapper from curses caused the curses module
to run, which attempted to import the _curses module, which did
not exist (as detailed above). The result was that an indirect
ModuleNotFoundError got raised whenever importing our decorators
module (which itself was found, as was the curses module in its
newly introduced spurious import statement, but the _curses module
was not, on Windows systems).

This bug was not spotted at the time it was introduced because it
was introduced while writing code to express the direction to pick
up at when resuming collaboration on the project.