greetall script doctests for missing and unopenable files
These tests are long and can later be moved to another file, which
I think we should do at the same time as we convert them to use
another testing framework (probably pytest, as it's good for IO).

I've made these now so that as make_greeter is implemented and
greet uses it, we can more easily and efficiently check that
greetall.py still appears to be working.