Revert "Revert "removed last and pick implementation bodies""
Re-removing last and pick implementation bodies. The revert of the
original removal remains on the ek/gc2-dev branch, while this
revert (i.e., re-removing the implementation bodies) is on the
ek/gc2 branch so that those two problems in gencomp1 can be worked
at the same time as the ones in gencomp2.

This reverts commit ec714d8c4602e8970204bfe43231840ba324373d.