pytest tests for _kronecker_delta & _identity_matrix_row
These satisfy the requirement I gave for identity_matrix_alt's
top-level helpers to be tested.

When I remove implementation bodies (leaving skeletons) to prepare
the tested exercises to be worked again, I intend to remove these
entirely, since they are part of what is to be done for the
exercise, rather than part of what specifies/checks the exercise.