More tests of dequeue ordering
The new tests test that the order in which a FIFO, LIFO, or PQ
dequeues items is in the correct order, for a small but nontrivial
number of items (significantly more than 3 together in a queue).

I've also done a tiny refactor in an existing test, so that tests
that build and compare lists all use the same lhs/rhs convention.