Fix my_product* naming convention and docstring descriptions
The doctests still need to be adjusted so the one that is expected
to be fast will take uncomfortably long if not fast enough, while
the one that is expected to be slow will *not* take uncomfortably
long at least if it is implemented correctly.

pytest tests are also still needed for these functions (among others).