Adjust my_product* tests for speed (docs still need work)
But I'm not sure I really should be calling them my_product and
my_product_slow. The slow one is much slower in this kind of test,
where only a tiny prefix of the full series of values is actually
generated, but I am not sure the slow version really is slower if
the input is going to be iterated through all the way.