Start on class versions of some gencomp functions
NB: More tests are needed, to ensure, for example, that iterators
implement __iter__ properly. It may also be good to add more
stringent requirements, including for the already-done function
versions, about when an iterator passed as an argument gets
consumed (usually it should not be consumed at all before the
iterator that wraps it begins to be consumed). Most of these tests,
as well as some of the existing tests, should go in a newly created
separate test module instead of being doctests. (Many tests should
be retained as doctests, but they, too, should be covered by
unittest or pytest tests in a separate test module).