Fix duplicate test class name (hiding tests) in test_functions.py
This fixes a bug where I meant to name the class following
TestAsIteratorLimited as TestAsIterator, but I wrongly also
called it TestAsIteratorLimited.

That caused only TestAsIterator tests to be run, and to be shown
as TestAsIteratorLimited (since the second class definition rebound
the name, causing the tests in the first class never to be found or
run). Now that TestAsIterator is correctly named, all tests run and
pass and, as far as I can tell, are always correctly shown and
classified in test runners.