Move scrapped exercises from recursion.py to subproblems_wip.py
This also edits subproblems_wip.ipynb accordingly, to change all
fully qualified referneces, both in code and in prose, to point to
the new locations.

It also has other minor changes to accompany this move, including
a couple comments explaining empty output when there should be one
node (because it's not worth fixing in the code in subproblems_wip
given that these exercises will not be used, and because there is
no prospect for confusion).