Fix ZipTwo/Zip next()-after-StopIteraiton side-effect bug
In ZipTwo, calling next after StopIteration that was raised due to
calling next on iterator2 would call next again on iterator1, which
is an observable state change that takes place even if there is no
bug in any other code. So this was objectively a bug in ZipTwo.

Likewise, in Zip, calling next after StopIteration that was raised
due to calling next on any but the first iterator would call next
again on all preceding iterators. So this was objective a bug in
Zip for the same reason.

This commit fixes both ZipTwo and Zip. Unfortunaely I don't yet
have tests that check for this bug, since the test_gencomp1.py
tests aren't yet written.