Lexicographical comparisons assume total ordering
This commit adds a notebook, basics/lexcomp.ipynb, intended to be
used broadly for examining lexicographical comparisons in Python,
but so far containing just one section, a writeup and demonstration
of how lexicographical order comparisons, even with the < and >
operators, use == to find a mismatch before any other comparison.

Accordingly, I've noted the wrong example section I had written at
the end of math/compare.ipynb as incorrect, with reference to the
information in basics/lexcomp.ipynb. This is so that one can figure
out what happened by examining the git history; I'll remove that
wrong section of compare.ipynb very soon, probably the next commit.