Test that distinct_dict_by_keys doesn't confuse keys
This adds a doctest, backported from 098a930. (Only the doctest is
backported and present in this commit, not any other changes. This
commit does not modify other code in distinct_dict_by_keys.)

This test checks that, when a subject key is present in one input
dict and absent in another, the implementation doesn't introduce
skew that causes another subject key to be confused with it (or
other keys to be confused with each other). But it is not
exhaustive. When unittest or pytest unit tests are added, it may be
good to add more tests for this.