Add pytest tests for gencomp1 enumerate, any/all, & zip exercises
The functions in gencomp1.py with new tests in test_gencomp1.py are:

* my_enumerate
* print_enumerated
* print_enumerated_alt
* my_any
* my_all
* my_zip
* print_zipped

zip_two tests were already present.

Other changes (which are related):

+ Add test that iter passes through unchanged in zip_two tests
+ Other small refinements in zip_two tests
+ Remove static type annotations and dynamic @typechecked checking

This is all in just test_gencomp1.py. This commit doesn't remove
any type annotations or @typechecked decorations from other test or
non-test modules.

The reason for the removal is that the annotations weren't very
extensive or useful in test_gencomp1.py, and this will make those
tests more similar to the way tests in test_gencomp2.py on the
ek/gc2-dev branch are currently written. The annotations that were
removed can easily be put back later. (Some annotations here could
be a bit tough to do well, but none of those were already present.)