Remove true but way too misleading misdirection
The anagrams exercise's description quantified a deliberate
overestimate on space usage and presented it accurately, but I
wrote it in such a way as to avoid giving away any solutions and to
even be deliberately misleading, but I think it was far *too*
misleading. So I've removed that part of the description.