Move tests for common iterator behavior to base class and add more
+ Use subtests for tests that made multiple assertions about
  multiple things (when they still did even after no longer
  checking that objects are iterators).

This is arguably less robust in some ways, in that if a function
that should always return an iterator returns one for some argument
values but not others, that might not be caught. But I don't think
all those checks are worthwhile; I think the disadvantage of the
longer, more complicated tests outweighs it.