Backport common iterator behavior test base class from 3faf65a
The class was originally test_gencomp2._CommonIteratorTests on the
ek/gc2-dev branch (introduced in commit 3faf65a). This commit
backports it to the ek/qa branch (which might sometime in the near
future itself be merged into ek/gc1-dev) and makes it public in
the new module file testing.py.