Remove older exercises from list in gencomp-gc0-notes.txt
They are currently the ek/gc0 exercises. The current plan is not
to merge any of the gencomp1 pytest tests, nor any of the new
exercises created on or backported to this branch, until those are
done.

If a branch of exercises is prepared from this branch (ek/qa)
before the current ek/gc0 exercies are done, then it would be
created from ek/gc0-dev and most likely be called ek/gc0-edge.