Refactor doctests to remove some import statements
This has the benefit of shortening various docstrings with no
reduction (and sometimes an increase) in test readability.

In some modules, it is okay for doctests to fail as a result of the
itertools module not being imported, since it's being substantially
used in those modules anyway (and if it were not used then the
intent of the exercises wouldn't be satisfied). Also, the style
we've been using has been to only occasionally use "from" imports
for things in itertools (so the "import itertools" is, and will
continue to be, present).