Do the bisector dependency injection FIXMEs
Except that I have kept all the FIXME text because the FIXMEs are
part of the exercises (rather than replacing them, as one would
normally do when working these exercises).

I've changed some wording and moved around some of the FIXMEs to
be in the places where the doctests should actually be added.