Backport review exercises from ek/search-dev (e23cfe2)
Further review exercises will be added to this branch. Although
the material so far was originally developed on the ek/search-dev
branch, it will be developed here, and commits here will be merged
to ek/search-dev (where search.py and possibly other related
exercises/material will continue to be developed).