Refactored _a_win_av gives different results from old version :(
So there is at least one bug in a least one of the versions.

I've kept both versions in the code (the old one should be removed
once the bug is found and fixed, even if it is in the old version,
so that they can be compared to ensure they give the same results).