Record initial position in history
This fixes one bug when the new _a_wins_av is used. This got taken
care of by the inverted logic the old one used, but must be done
explicitly now. But the fixed new code still produces different
results, so something *else* is wrong with at least one version.