Workaround for bug in old logic
I found the bug in the old _a_wins_av_old logic: whereas B's start
coordinates are added to the original b.vis on the top-level call,
A's start coordinates are not added to the original a.vis until A
revisits that spot the first time.

This comment applies a workaround. Both versions now appear to
behave the same, including on all tests.