Turn count_av_b_wins into count_av_a_wins
My claim that A was favored was incorrect (it arose from analysis
of results using the old _a_wins_av_old logic that had a bug).

Since B is actually favored, it makes more sense to report details
of, and count, A's wins. (But the game does not appear as
unbalanced in favor of B as I had thought it was unbalanced in
favor of A.)