Remove a confusing subexercise
One of the reasons it is confusing is that there seems to be some
disagreement among computer scientists about how, exactly, the
asymptotic notation should be defined for multivariable functions.
Is it okay for all variables to need to be sufficiently large? Or
does the inequality need to hold whenever at least one variable is
sufficiently large? The latter seems most common (though it is
often cited to a definition in a CLRS exercise that may or may not
be intended as an authoritative statement of CS practice). But in
looking into this, I've seen the former used as well. That affects
which solution to the subexercise (that this commit removes) would
be considered correct.

Alternatively, if I am mistaken and one definition is decisively
preferred, then while such (sub)exercises could potentially be
reasonable, I don't want to include any at this point, at least
until my own confusion is cleared up, without which I cannot judge
how to write them (or... what the correct answers would be).