Give more information in two_sum_int description
But is it too much?

I don't want to give the exercise away, but I also don't want the
description to lead people to implement a variant of radix sort,
so a hint about a specific way the intended solution is similar to
(the intended solution to) two_sum_fast may be needed.