Make two_sum_slow_alt, initial with a comprehension
This is not (at least not currently) intended as its own exercise,
just an alternative implementation for testing the two_sum_slow
exercise.

This implementation is documented as using a loop (without nesting)
but actually uses a comprehension. I'm about to change it to a
loop, which in this case I think is clearer.