Remove a FIXME posing an insufficiently useful subexercise
The removed FIXME was mostly misdirection: it said if something
wasn't done in previous FIXMEs, then to do something else, but it
was someting that should definitely have been done in some of the
previous FIXMEs, so the description of the other thing to do was
all smoke and mirrors.

I don't think that's inherently bad, but the description of the
thing that would never actually be done was too long, and not
interesting enough, to be worth reading and thinking about.