Add "alt" drawings separately in object_graph.ipynb
This is to facilitate easy visual comparison between the output of
draw_tuples (including draw_one_tuple, which delegates to it) and
draw_tuples_alt. There should be (and are) no differences, when
they are passed the same objects (in the same order) as arguments.

Interpreted as plaintext (as Git does), the number of lines in the
notebook decreased because I happened to be working on it in a
Jupyter client that saves embedded SVG images "minified," thereby
dropping newlines.