Remove a note about a merge that probably won't be done
I'm no longer planning to merge the ek/drawing-dev problems
directly to ek/review-dev to be presented along with them and at
the same time. I think it's better to consider them separate
problem sets and to make it easy to merge the "review" exercises to
master before (or, I think less likely, after) the "drawing" ones.