More on possible give_metadata_from augmentation exercise
I don't plan to include this at this time. It's too complicated to
be a good choice of first exercise in this problem set, and also
there is a benefit to having a very simple decorator factory that
works like @functools.wraps in the most important ways, which this
does.