Add "support" functions identity_function and make_fmap
These are not exercises, and I do not plan to remove or alter their
implementations when preparing exercises to be worked. Instead they
help demonstrate what some exercises' code should do and why it is
meaningful or interesting to do it.

Currently these are just demonstrating @wrap_callable_args, which
is all I intend make_fmap to do, but identity_function, being quite
simple, yet preferable to lambda x: x in a variety of situations,
will likely get more use.