Describe and implement possible @subscribe exercise
I don't think I'll keep this for this problem set, though it may be
a good exercise later. It's too design-heavy for such a big problem
set that is not focused on these particular design issues, in this
position in decorators.py where it is intended as one of the "easy"
exercises in the problem set.