Make test_decorators.txt work with pytest by itself
That is, make this work even when no .py files are being imported
to load tests fromt. Since the code is not (curently) packaged, the
import must find decorators.py in the project directory, but pytest
does not add the current directory to sys.path when there is no
module being loaded to find/run tests from.