Allow StopIteration propagation in as_closeable_func test
This passes the closeable function constructed from an iterator to
as_iterator_alt, rather than as_iterator, to convert it back into
an iterator. This is necesary because we want StopIteration to
propagate up rather than be converted to RuntimeError (and we don't
want to modify as_iterator to catch StopIteration and return,
because that wouldn't always be the right thing to do, and one of
the benefits of generators is that they prohibit this).

The reason this change is necessary is that, during the course of a
recent review reworking of the problems in functions.py on the
master branch, as_iterator ended up as the generator version and
as_iterator_alt ended up as the 2-arg iter call version, rather
than the other way around as it previously had been.

It might be good to comment or even specify requirements as to
which is which, to improve the reusability of these exercises in
other settings (students might not be comfortable troubleshooting,
judging if they should modify, and, where reasonable, modifying
existing doctests that are already written to evaluate their work's
correctness). But I'm holding off on that for now, since the
implementations might be switched again: in particular, as_iterator
and as_iterator_limited should probably either both use the iter
builtin or neither.