Test MonkeyPatch reuse, including nested decorations and calls
These tests are (as documented in their names) testing that the
same MonkeyPatch instances can be used multiple times when they are
used as decorators (even in nested decorated definitions and/or
nested wrapper calls).

The tests do not attempt to test reuse as a context manager, or use
of the same instance as both a context manager and a decorator,
because although it is possible to implement and guarantee those
behaviors, I intentionally did not document them as a requirement
for MonkeyPatch (and my own implementation neither attempts to
guarantee, nor always satisfies, such a requirement).

Maybe a later version of the exercise will also require that, but I
think the exercise is demanding/complicated enough for now.