Move _*TREE_FACTORIES out of TestStructuralEqual
So they can be used by other classes.

Most classes, at least so far, cannot make good use of them in the
current design, since for most of the code under test, expected
results cannot be inferred from them. But sometimes it can and,
when it can, there is no benefit to repeating the same exact test
method body numerous times in the same class.