List factories; make ill-considered special-purpose parameterizers
The newly listed factories will be useful and should be kept.

The parameterizers were intended just for TestReflectInPlace, or
possibly for it and a small number of related classes. This is not
a good idea, though. It's not clear that the complexity of this
design couldn't be avoided even if I needed to do what I thought I
needed to do, but also, for in-place reflection, the tree has to be
mutable, and there is currently no plan to have multiple mutable
node types. (We may have them eventually, but even if so, I cannot
predict whether they should or shouldn't be tested in this way if
we do.) Because of that, the parameterization can be simpler, and
either shouldn't need special-purpose parameterization functions at
all or, if it does, shouldn't need anything compared to the
complexity of the one here.