Remove @for_implementations TODO
It looks like this may be our only sizable test module where we
are using @parameterized_class in a nontrivial yet readily
understandable way. It's valuable to have such a thing in the
project. Furthermore, the faciltity this TODO was for was too
general and involved to really belong in test_functions.py, but
insufficiently general and elegant in its design to be a good
general-purpose facility.

Further improvement could be in a direction opposite that of the
TODO removed here. That is, using @parameterized_class in a more
demonstrative and idiomatic way than currently.