Finish TestBinarySearch
By writing the "medium"-sized tests.

Note that there are still more tests for tree.binary_search* to be
written, since, as stated in the TestBinarySearch class docstring,
tests that binary_search and binary_search_bst return the same
results when searching the same tree for the same value) will go in
a separate test class, which is not yet written.