Put UnsortedFlatTable before SortedFlatTable
This is mainly becuase I'm planning to require SortedFlatTable to
be reversible, including its three mapping views being reversible,
which tips the scales toward it being more challenging than
UnsortedFlatTable.

But it is also arguably a benefit that it lets SortedFlatTable come
immediately before BinarySearchTree (whereas UnsortedFlatTable will
be even farther away from HashTable, but this is less of an issue
since DirectAddressTable, which still comes immediately before
HashTable, reviews the important ideas).