Make Vec handle NaNs as list/tuple do + test this
I had originally planned to omit this requirement, but it is not
reasonable to omit it, because whatever behavior is implemented
should be consistent between __eq__, index, count, remove, etc.,
and the other methods besides __eq__ already have mixins inherited
from Sequence that provide the behavior of comparing with "is" as
well as with "==".

Having this requirement for sequences.Vec has the further advantage
that it is less new material that may need to be learned at once
while doing working the mappings exercises I intend to appear later
in this same problem set.