Make _ordered_equality a decorator factory
It now takes a key equivalence predicate and returns a decorator
that defines equality of mapping instances that is ordered in the
sense of taking item iteration order into account.

This decouples the two separate sense of "ordered equality" for
mappings: whether the order in which items are yielded is taken
into account, and whether the key equivalence predicate itself
determines equivalence using a technique that is based on equality
comparison or on order comparison.

Currently only DirectAddressTable uses _ordered_equality with
_eq_equivalent; other classes that use it use _le_equivalent. If
ordered mappings, in the sense of OrderedDict, are added in the
future, they might also have @_ordered_equality(_eq_equivalent).