Fill in MonkeyPatchAlt with pre-decorator MonkeyPatch code
This is so only the decorator-specific code has to added.

The code filled in is from earlier in the master branch (not from
when MonkeyPatch was originally written). It had last appeared in
this form (but in MonkeyPatch instead of MonkeyPatchAlt) in
0beae6b (though that commit was not related to it). Except I've
added a specific FIXME about making it work as a decorator without
changing how it works as a context manager.