Extract _collect_if_not_ref_counting for reuse
This backports a change from the ek/drawing-dev branch, moving
test_functions._collect_if_not_ref_counting to the (on this branch,
newly created) testing module as collect_if_not_ref_counting.

The rest of the contents of testing.py from ek/drawing-dev are not
backported here. This is not yet used by multiple modules on the
master branch (contrary to the testing module's docstring) but it
soon will be.