Support reentrant sll.Node lookup/construction
This happens when a hash computation, occurring in the check for a
an existing node with a given value and next node, itself creates
(or attempts to create) a node.

test_sll.CachedEq is a non-contrived example of when it would not
merely be technically permitted, but actually reasonable, for code
to do this. However, we may ultimately choose not to support this.

The implementation of sll.Node is "fixed" so this doesn't deadlock,
but the current "fix" shall soon be changed, as it causes more
subtle problems, since __eq__, if called during table insertion,
could cause sll.Node.__new__ to be reentered to create a node it
has already created but not yet finished installed in the table.
I plan to expand the CachedEq docstring and/or adjacent materials
to cover this issue. (Failing that, I would revert this commit and
leave the deadlock, which is still better than allowing sll.Node to
violate its invariants even when other code does nothing "wrong.")