Show the heterogeneous cycle problem is unsolved
Given the complexity, it may not be worthwhile to get this working.
Even if I do, I suspect the techniques will have major
disadvantages, and also be excessively complicated for use in
demonstrations and as exercises. It may be better to simply note
that the implementation can leak heterogeneous cycles through the
WeakValueDictionary if elements hold mutable state that doesn't
prevent them from being immutable in the relevant sense (due to not
contribute to their values, per equality comparison and hashing)
and that state maintains strong references back to node objects.