Rename Node to HashNode; remove ImportantPoint
This renames sll.Node to sll.HashNode, updating code, docstrings
including doctests, and the tests in test_sll.py. It also rewords
comments and docstrings accordingly, plus some addititional editing
for clairy.

This is different from what I planned to do, as written in the FIXME
that had appeared under sll.py's module docstring. I think naming
the node type HashNode is sufficient to clarfy that it's not a plain
node, and that the comments and docstrings are sufficient to make
clear that this hash consing implementation is not suitable for
production use, at least in a general setting where heterogeneous
cycles may occur.

I have also removed test_sll.ImportantPoint. This can be made to
stop interfering with other tests by writng cleanup code, but I
don't think this class is needed anymore, due to the way I've
reminded the reader of the sll module docstring as well as various
other docstrings of the relevant issues. Jamming it in at the end
to make up for insufficient clarity about which sources of wrong
behavior the node type is responsible for preventing was not ideal,
anyway.