Shorten sll.py doctests; put big ones in test_sll.txt
This creates a new test file, test_sll.txt, to hold the original
long sll.py doctests for sll.HashNode and sll.traverse, with slight
adaptations to that format. The versions in sll.py are then
considerably shortened, with the hope that this will make them
easier to read as documentation, and that it will make the code
a bit nicer to work with.