Reorganize HashNode tests; rework docstrings
This keeps the portion of the old sll module docstring that remains
useful in light what appears (though I may want to add more tests)
to be a correct solution to the heterogeneous cycle problem that,
while not obvious, seems straightforward enough for presentation
and possibly for inclusion in a guided exercise. This is still very
much a draft.