Remove non-slotted version of "manual data class"
This is to keep down the explosion of variations appearing as
demonstrations or exercises in aggregate.py. The point of classes
with explicit dunders in this module is to motivate why named
tuples and data classes are useful, not to showcase the details of
how one would implement every possibly useful data-class-like
behavior with them. If I kept this class, then it would be very
misleading not also to have a non-slotted immutable version (which
I had originally planned, too), and moderately misleading not to
address the issue of whether additional attributes can be added to
each version and, for the mutable version, if it allows them,
whether they should participate in equality comparison and hashing.

This also removes slotted versions' tests for __dict__ absence. Not
all desired behaviors are tested in these doctests, and that is not
one of the more important behaviors to doctest, in those cases. (In
contrast, for named tuples, once added to this module, that's
another story, and the doctests for those should show the absence
of instance dictionaries.)