Move data classes materia to new notebook
One of the changes in a36a608 was to remove data_classes.ipynb and
put that material at the bottom of classes3.ipynb. This undoes that
specific change (since I anticipate classes3.ipynb will increase in
length to the point where it should go in another notebook, but
also because I don't intend to express that all the other material
in classes3.ipynb needs to be done first).

The file is called aggregate.ipynb instead of data_classes.ipynb.