Refactor restructure logic to one script file; add help
This combines all three files that contained bash code to help with
the forthcoming project restructure into a single file, yet overall
modularity is improved. A top comment is added, and an informative
usage message is added. Most code lives in several named functions.

Because there is just one script, performing the tasks it is asked
to do via command-line argument, it is no longer necessary to have
a separate "check" script; just run "shellcheck meta/restructure".

Although prior functionality should be intact, patching imports in
Jupyter notebook is not yet implemented. One way to look at these
changes is as preparation for adding that feature.