Fix meta/restructure regex bugs
Parentheses were absent around something that needed to be captured
(which was not caught because the pattern still usually worked).

Also, \w was used inside [], but ERE does not recognize that.

This commit fixes those bugs, generalizes one of the regexes so
both kinds of imports allow dots, and reorganized/refactors the
code so it easier to understand, so such bugs don't slip back in.
(This refactor may also help in writing patch_notebook_imports.)