Stop using separate "qa" conda environment
Other branches differ almost as much as this in the packages they
specify in environnment.yml, but are not using separate envs for
that reason. At this point, the ek/qa branch exists mainly to be
merged into gencomp-related branches, rather than to be maintained
as an independent branch parallel to master but with substantially
different tooling and tests.