Remove test_greetall.py, test_hola.py, and some tools, for now
This is how the ek/pytest-dev branch diverges from ek/qa. The
material in ek/qa that will be ready to merge sooner is the tests
in test_gencomp1.py, which this commit keeps.

In the future, this commit may be partially or fully reverted, but
the removed material is not useful at the moment.