Backport some gencomp2 tests from ek/gc2-dev (8c097a8) to ek/gc1-dev
This doesn't actually involve adding new test logic, just adding
additional rows (parameters) for already parameterized test methods
and updating docstrings and comments. But it is a backport from
ek/gc2-dev in the sense that the most important changes--the rows
for @pytest.mark.parameterize--are already in the ek/gc2-dev tests.