Eliminate requirements.txt; use pyproject-flake8
Before, requirements.txt had a single optional dependency that was
not available from a conda repository, doctestfn. That has been
promoted to be on the same "level" as the other testing-related
dependencies, specified as a dependency that must be installed
through pip in environment.yml (plus normally in pyproject.toml).

In 7ebb6b7, I had gone back to using flake8 and having a separate
.flake8 configuration file, rather than using pyproject-flake8,
which must be run as "pflake8" rather than "flake8". Now, I have
added Flake8-pyproject, which is also called flake8-pyproject but
should not be confused with pyproject-flake8. Flake8-pyproject
monkeypatches flake8 to read configuration from pyproject.toml.
This project's poetry (pyproject.toml) and conda (environment.yml)
dependencies now include it. Since it's not (or not yet) available
in a conda repository, it is specified as a dependency that must be
installed through pip in environment.yml, which specifies flake8
separately so that it and its dependencies can still be installed
from conda-forge.