License project 0BSD; improve docs and metadata
More information about licensing should be added. The license
should be named and linked in the top-level README.md, and it
would be best to note the license (with SPDX-License-Identifier) in
each Jupyter notebook.

There are steps for making the license clear, detailed in a FIXME
comment in the top-level README.md. Some of those are done in this
commit, while others remain to be done.

In addition, if possible, some details of the top-level README
ought to be broken out into one or more separae files in doc/, and
more information added in doc/. That, too, is detailed in comments
in the top-level README.