Keep ipython at 8.8.* for now
This commit also updates unrelated dependencies.

IPython 8.9.0 changes the interface for autosuggestions. Among
other differences, the right-arrow key completes by a single
character instead of the entire suggested line:

https://ipython.readthedocs.io/en/stable/whatsnew/version8.html#ipython-8-9-0

The end key continues to work to complete the entire suggestion.

It seems the keys aren't configurable, though that is planned:

   "[W]e are also working on making the shortcuts configurable."

The new autosuggestion interface includes significant improvements.
(For example, completing the whole suggestion shows a suggestion
for how the line can be further extended, if there is one.) Also,
the new keys are arguably more logical and intuitive, in that they
resemble the way the corresponding keys work in most editors.

But I think there are two weaknesses that justify holding off a bit
in taking this update:

1. The popular shells with analogous features, fish and pwsh, both
   use right-arrow to accept the whole line.

2. On many keyboards, it is a further reach to the end key than the
   right-arrow key.

I consider at least the first to be a strong consideration.

Once this becomes configurable, there will be no reason to keep
ipython back. We should also accept updates if needed to fix any
major bugs or any security bugs (of any severity). Otherwise, I
suggest we wait a short while to see how these features evolve.