Don't update pipx packages in dev container
Permissions are not set to allow it, at least in the case of
poetry, installed via pipx but through a feature. I don't want to
elevate privileges to do it, and I'm not sure it's a good idea in
general, really.

If we get a sufficiently recent version of poetry (and differences
to the latest stable are not security-related), then it's fine to
use a slightly older version.