Possible workaround for conda bug (vscode pytest)
palgoviz (not specifically the ek/container branch) is affected by
https://github.com/conda/conda/issues/12082.

The effect in palgoviz is that whether pytest discovery works in VS
Code seems to be determined by details of how conda is installed.
In practice, test discovery seems always to work on Windows but only
occasionally on GNU/Linux (even though that does not appear to be
the general pattern of the conda bug).

See also:

- https://github.com/microsoft/vscode-python/issues/19069
- https://github.com/microsoft/vscode-python/issues/18493

This bug should be worked around in palgoviz because, currently,
the conda instructions in README.md and doc/install-with-conda.md
will likely fail to work for a substantial fraction of users.

We depend on conda-build for "conda develop ." or "mamba develop ."
but "pip install -e ." should meet our needs (though I feel it is
less elegant). This commit removes the conda-build dependency and
runs "pip install -e ." in the dev container.

NOTE: These are NOT the only changes that need to be made. If we
are going use this workaround, then those two documentation files
must also be updated to recommend it.